+++ lib/elixir/test/elixir/list_test.exs
@@ -49,12 +49,6 @@ defmodule ListTest do
     assert List.foldr([1,2,3,4], 0, fn x,y -> x - y end) == -2
   end
 
-  test :member? do
-    assert List.member? [1,2,3], 1
-    refute List.member? [1,2,3], 0
-    refute List.member? [], 0
-  end
-
   test :concat_1 do
     assert List.concat([[1,[2],3], [4], [5,6]]) == [1,[2],3,4,5,6]
   end